+++ src/libtrx/game/ui/dialogs/graphic_settings.c
@@ -4,27 +4,28 @@
 #include "strings.h"
 
 static const UI_SETTINGS_ENUM_ENTRY m_HealthBarShowModeEnumEntries[] = {
-    { BSM_DEFAULT, GS_ID(GRAPHIC_SETTINGS_DEFAULT) },
-    { BSM_FLASHING_OR_DEFAULT, GS_ID(GRAPHIC_SETTINGS_FLASHING_OR_DEFAULT) },
-    { BSM_FLASHING_ONLY, GS_ID(GRAPHIC_SETTINGS_FLASHING_ONLY) },
-    { BSM_ALWAYS, GS_ID(GRAPHIC_SETTINGS_ALWAYS) },
-    { BSM_NEVER, GS_ID(GRAPHIC_SETTINGS_NEVER) },
-    { BSM_PS1, GS_ID(GRAPHIC_SETTINGS_PS1) },
+    { BSM_DEFAULT, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_DEFAULT) },
+    { BSM_PS1, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_PS1) },
+    { BSM_FLASHING_OR_DEFAULT,
+      GS_ID(GRAPHIC_SETTINGS_BAR_MODE_FLASHING_OR_DEFAULT) },
+    { BSM_FLASHING_ONLY, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_FLASHING_ONLY) },
+    { BSM_ALWAYS, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_ALWAYS) },
+    { BSM_NEVER, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_NEVER) },
     { -1, nullptr },
 };
 
 static const UI_SETTINGS_ENUM_ENTRY m_AirBarShowModeEnumEntries[] = {
-    { BSM_DEFAULT, GS_ID(GRAPHIC_SETTINGS_DEFAULT) },
-    { BSM_FLASHING_ONLY, GS_ID(GRAPHIC_SETTINGS_FLASHING_ONLY) },
-    { BSM_NEVER, GS_ID(GRAPHIC_SETTINGS_NEVER) },
-    { BSM_PS1, GS_ID(GRAPHIC_SETTINGS_PS1) },
+    { BSM_DEFAULT, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_DEFAULT) },
+    { BSM_PS1, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_PS1) },
+    { BSM_FLASHING_ONLY, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_FLASHING_ONLY) },
+    { BSM_NEVER, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_NEVER) },
     { -1, nullptr },
 };
 
 static const UI_SETTINGS_ENUM_ENTRY m_EnemyHealthBarShowModeEnumEntries[] = {
-    { BSM_ALWAYS, GS_ID(GRAPHIC_SETTINGS_ALWAYS) },
-    { BSM_NEVER, GS_ID(GRAPHIC_SETTINGS_NEVER) },
-    { BSM_BOSS_ONLY, GS_ID(GRAPHIC_SETTINGS_BOSS_ONLY) },
+    { BSM_ALWAYS, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_ALWAYS) },
+    { BSM_NEVER, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_NEVER) },
+    { BSM_BOSS_ONLY, GS_ID(GRAPHIC_SETTINGS_BAR_MODE_BOSS_ONLY) },
     { -1, nullptr },
 };
 
@@ -183,6 +184,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .option_type = COT_INT32,
         .label_id = GS_ID(GRAPHIC_SETTINGS_FOG_START),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_FOG_START_DESCRIPTION),
         .target = &g_Config.visuals.fog_start,
         .min_value = 1,
         .max_value = 100,
@@ -191,6 +193,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .option_type = COT_INT32,
         .label_id = GS_ID(GRAPHIC_SETTINGS_FOG_END),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_FOG_END_DESCRIPTION),
         .target = &g_Config.visuals.fog_end,
         .min_value = 1,
         .max_value = 100,
@@ -199,6 +202,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .option_type = COT_RGB888,
         .label_id = GS_ID(GRAPHIC_SETTINGS_WATER_COLOR_R),
+        .description_id = nullptr,
         .target = &g_Config.visuals.water_color,
         .min_value = 0,
         .max_value = 255,
@@ -210,6 +214,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .option_type = COT_RGB888,
         .label_id = GS_ID(GRAPHIC_SETTINGS_WATER_COLOR_G),
+        .description_id = nullptr,
         .target = &g_Config.visuals.water_color,
         .min_value = 0,
         .max_value = 255,
@@ -221,6 +226,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .option_type = COT_RGB888,
         .label_id = GS_ID(GRAPHIC_SETTINGS_WATER_COLOR_B),
+        .description_id = nullptr,
         .target = &g_Config.visuals.water_color,
         .min_value = 0,
         .max_value = 255,
@@ -232,6 +238,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .option_type = COT_ENUM,
         .label_id = GS_ID(GRAPHIC_SETTINGS_CAMERA_MODE),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_CAMERA_MODE_DESCRIPTION),
         .target = &g_Config.visuals.camera_mode,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -242,6 +249,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .option_type = COT_INT32,
         .label_id = GS_ID(GRAPHIC_SETTINGS_FOV),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_FOV_DESCRIPTION),
         .target = &g_Config.visuals.fov_value,
         .min_value = 30,
         .max_value = 150,
@@ -252,12 +260,14 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .target = &g_Config.visuals.fov_vertical,
         .label_id = GS_ID(GRAPHIC_SETTINGS_FOV_VERTICAL),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_FOV_VERTICAL_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 #elif TR_VERSION == 2
     {
         .option_type = COT_INT32,
         .label_id = GS_ID(GRAPHIC_SETTINGS_FOV),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_FOV_DESCRIPTION),
         .target = &g_Config.visuals.fov,
         .min_value = 30,
         .max_value = 150,
@@ -268,6 +278,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .option_type = COT_BOOL,
         .label_id = GS_ID(GRAPHIC_SETTINGS_USE_PSX_FOV),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_USE_PSX_FOV_DESCRIPTION),
         .target = &g_Config.visuals.use_psx_fov,
     },
 #endif
@@ -276,12 +287,14 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .option_type = COT_BOOL,
         .label_id = GS_ID(GRAPHIC_SETTINGS_REFLECTIONS),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_REFLECTIONS_DESCRIPTION),
         .target = &g_Config.visuals.enable_reflections,
     },
 
     {
         .target = &g_Config.visuals.enable_skybox,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_SKYBOX),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_SKYBOX_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 #endif
@@ -289,18 +302,21 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .target = &g_Config.visuals.enable_braid,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_BRAID),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_BRAID_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 
     {
         .target = &g_Config.visuals.enable_breeze,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_BREEZE),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_BREEZE_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 
     {
         .target = &g_Config.visuals.enable_3d_pickups,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_3D_PICKUPS),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_3D_PICKUPS_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 
@@ -308,18 +324,24 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .target = &g_Config.gameplay.enable_pickup_aids,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_PICKUP_AIDS),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENABLE_PICKUP_AIDS_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 
     {
         .target = &g_Config.visuals.enable_ps1_crystals,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_PS1_CRYSTALS),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENABLE_PS1_CRYSTALS_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 
     {
         .target = &g_Config.visuals.enable_round_shadow,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_ROUND_SHADOW),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENABLE_ROUND_SHADOW_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 #endif
@@ -327,6 +349,8 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .target = &g_Config.visuals.enable_gun_lighting,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_GUN_LIGHTING),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENABLE_GUN_LIGHTING_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 
@@ -334,6 +358,8 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .target = &g_Config.visuals.enable_shotgun_flash,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_SHOTGUN_FLASH),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENABLE_SHOTGUN_FLASH_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 #endif
@@ -347,6 +373,7 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
     {
         .option_type = COT_DOUBLE,
         .label_id = GS_ID(GRAPHIC_SETTINGS_UI_TEXT_SCALE),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_UI_TEXT_SCALE_DESCRIPTION),
         .target = &g_Config.ui.text_scale,
         .min_value = 50,
         .max_value = 200,
@@ -357,6 +384,7 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
     {
         .option_type = COT_DOUBLE,
         .label_id = GS_ID(GRAPHIC_SETTINGS_UI_BAR_SCALE),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_UI_BAR_SCALE_DESCRIPTION),
         .target = &g_Config.ui.bar_scale,
         .min_value = 50,
         .max_value = 200,
@@ -367,6 +395,8 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
     {
         .option_type = COT_BOOL,
         .label_id = GS_ID(GRAPHIC_SETTINGS_UI_SCROLL_WRAPAROUND),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_UI_SCROLL_WRAPAROUND_DESCRIPTION),
         .target = &g_Config.ui.enable_wraparound,
     },
 
@@ -374,6 +404,7 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
     {
         .target = &g_Config.ui.menu_style,
         .label_id = GS_ID(GRAPHIC_SETTINGS_MENU_STYLE),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_MENU_STYLE_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_MenuStyleEnumEntries,
     },
@@ -382,18 +413,23 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
     {
         .target = &g_Config.ui.lara_health_bar.show_mode,
         .label_id = GS_ID(GRAPHIC_SETTINGS_HEALTHBAR_SHOW_MODE),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_HEALTHBAR_SHOW_MODE_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_HealthBarShowModeEnumEntries,
     },
     {
         .target = &g_Config.ui.lara_health_bar.location,
         .label_id = GS_ID(GRAPHIC_SETTINGS_HEALTHBAR_LOCATION),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_HEALTHBAR_LOCATION_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_BarLocationEnumEntries,
     },
     {
         .target = &g_Config.ui.lara_health_bar.color,
         .label_id = GS_ID(GRAPHIC_SETTINGS_HEALTHBAR_COLOR),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_HEALTHBAR_COLOR_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_BarColorEnumEntries,
     },
@@ -401,18 +437,21 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
     {
         .target = &g_Config.ui.lara_air_bar.show_mode,
         .label_id = GS_ID(GRAPHIC_SETTINGS_AIRBAR_SHOW_MODE),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_AIRBAR_SHOW_MODE_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_AirBarShowModeEnumEntries,
     },
     {
         .target = &g_Config.ui.lara_air_bar.location,
         .label_id = GS_ID(GRAPHIC_SETTINGS_AIRBAR_LOCATION),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_AIRBAR_LOCATION_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_BarLocationEnumEntries,
     },
     {
         .target = &g_Config.ui.lara_air_bar.color,
         .label_id = GS_ID(GRAPHIC_SETTINGS_AIRBAR_COLOR),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_AIRBAR_COLOR_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_BarColorEnumEntries,
     },
@@ -420,18 +459,24 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
     {
         .target = &g_Config.ui.enemy_health_bar.show_mode,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENEMY_HEALTHBAR_SHOW_MODE),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENEMY_HEALTHBAR_SHOW_MODE_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_EnemyHealthBarShowModeEnumEntries,
     },
     {
         .target = &g_Config.ui.enemy_health_bar.location,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENEMY_HEALTHBAR_LOCATION),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENEMY_HEALTHBAR_LOCATION_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_BarLocationEnumEntries,
     },
     {
         .target = &g_Config.ui.enemy_health_bar.color,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENEMY_HEALTHBAR_COLOR),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENEMY_HEALTHBAR_COLOR_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_BarColorEnumEntries,
     },
@@ -440,6 +485,8 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
     {
         .target = &g_Config.ui.enable_smooth_bars,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_SMOOTH_BARS),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENABLE_SMOOTH_BARS_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 #endif
@@ -447,12 +494,16 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
     {
         .target = &g_Config.visuals.enable_fade_effects,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_FADE_EFFECTS),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENABLE_FADE_EFFECTS_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 
     {
         .target = &g_Config.visuals.enable_exit_fade_effects,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ENABLE_EXIT_FADE_EFFECTS),
+        .description_id =
+            GS_ID(GRAPHIC_SETTINGS_ENABLE_EXIT_FADE_EFFECTS_DESCRIPTION),
         .option_type = COT_BOOL,
     },
 
@@ -465,6 +516,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_INT32,
         .label_id = GS_ID(GRAPHIC_SETTINGS_FPS),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_FPS_DESCRIPTION),
         .target = &g_Config.rendering.fps,
         .min_value = 30,
         .max_value = 60,
@@ -475,6 +527,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_ENUM,
         .label_id = GS_ID(GRAPHIC_SETTINGS_TEXTURE_FILTER),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_TEXTURE_FILTER_DESCRIPTION),
         .target = &g_Config.rendering.texture_filter,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -484,6 +537,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_BOOL,
         .label_id = GS_ID(GRAPHIC_SETTINGS_TRAPEZOID_FILTER),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_TRAPEZOID_FILTER_DESCRIPTION),
         .target = &g_Config.rendering.enable_trapezoid_filter,
     },
 
@@ -492,6 +546,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .target = &g_Config.rendering.anisotropy_filter,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ANISOTROPY_FILTER),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_ANISOTROPY_FILTER_DESCRIPTION),
         .option_type = COT_FLOAT,
         .min_value = 1 * 100,
         .max_value = 32 * 100,
@@ -502,6 +557,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_ENUM,
         .label_id = GS_ID(GRAPHIC_SETTINGS_FBO_FILTER),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_FBO_FILTER_DESCRIPTION),
         .target = &g_Config.rendering.fbo_filter,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -510,6 +566,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
 
     {
         .label_id = GS_ID(GRAPHIC_SETTINGS_RESOLUTION),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_RESOLUTION_DESCRIPTION),
         .custom_handler = {
             .format_value = M_ScreenResolution_FormatValue,
             .can_change_value = M_ScreenResolution_CanChangeValue,
@@ -520,6 +577,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_ENUM,
         .label_id = GS_ID(GRAPHIC_SETTINGS_RENDER_MODE),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_RENDER_MODE_DESCRIPTION),
         .target = &g_Config.rendering.render_mode,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -529,12 +587,14 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_BOOL,
         .label_id = GS_ID(GRAPHIC_SETTINGS_VSYNC),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_VSYNC_DESCRIPTION),
         .target = &g_Config.rendering.enable_vsync,
     },
 
     {
         .option_type = COT_FLOAT,
         .label_id = GS_ID(GRAPHIC_SETTINGS_BRIGHTNESS),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_BRIGHTNESS_DESCRIPTION),
         .target = &g_Config.visuals.brightness,
         .min_value = CONFIG_MIN_BRIGHTNESS * 100,
         .max_value = CONFIG_MAX_BRIGHTNESS * 100,
@@ -546,12 +606,14 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_BOOL,
         .label_id = GS_ID(GRAPHIC_SETTINGS_DEPTH_BUFFER),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_DEPTH_BUFFER_DESCRIPTION),
         .target = &g_Config.rendering.enable_zbuffer,
     },
 
     {
         .option_type = COT_ENUM,
         .label_id = GS_ID(GRAPHIC_SETTINGS_LIGHTING_CONTRAST),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_LIGHTING_CONTRAST_DESCRIPTION),
         .target = &g_Config.rendering.lighting_contrast,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -561,6 +623,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_ENUM,
         .label_id = GS_ID(GRAPHIC_SETTINGS_RENDER_MODE),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_RENDER_MODE_DESCRIPTION),
         .target = &g_Config.rendering.render_mode,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -570,6 +633,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_ENUM,
         .label_id = GS_ID(GRAPHIC_SETTINGS_ASPECT_MODE),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_ASPECT_MODE_DESCRIPTION),
         .target = &g_Config.rendering.aspect_mode,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -579,6 +643,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_INT32,
         .label_id = GS_ID(GRAPHIC_SETTINGS_SCALER),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_SCALER_DESCRIPTION),
         .target = &g_Config.rendering.scaler,
         .min_value = 1,
         .max_value = 4,
@@ -589,6 +654,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_FLOAT,
         .label_id = GS_ID(GRAPHIC_SETTINGS_SIZER),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_SIZER_DESCRIPTION),
         .target = &g_Config.rendering.sizer,
         .min_value = 40,
         .max_value = 200,
@@ -600,6 +666,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .target = &g_Config.rendering.screenshot_format,
         .label_id = GS_ID(GRAPHIC_SETTINGS_SCREENSHOT_FORMAT),
+        .description_id = GS_ID(GRAPHIC_SETTINGS_SCREENSHOT_FORMAT_DESCRIPTION),
         .option_type = COT_ENUM,
         .misc = &m_ScreenshotFormatEnumEntries,
     },